pull request feedback.  Moving node, job, task calls out of the ClusterStoreService and putting it in a ClusterStore class that is able to see the full store.
