+++ server/src/main/java/com/continuuity/loom/scheduler/JobScheduler.java
@@ -15,7 +15,6 @@
  */
 package com.continuuity.loom.scheduler;
 
-import com.continuuity.loom.account.Account;
 import com.continuuity.loom.cluster.Cluster;
 import com.continuuity.loom.cluster.Node;
 import com.continuuity.loom.codec.json.JsonSerde;
@@ -31,8 +30,8 @@ import com.continuuity.loom.scheduler.task.JobId;
 import com.continuuity.loom.scheduler.task.TaskConfig;
 import com.continuuity.loom.scheduler.task.TaskId;
 import com.continuuity.loom.scheduler.task.TaskService;
+import com.continuuity.loom.store.cluster.ClusterStore;
 import com.continuuity.loom.store.cluster.ClusterStoreService;
-import com.continuuity.loom.store.cluster.ClusterStoreView;
 import com.google.common.base.Function;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Iterables;
@@ -62,8 +61,7 @@ public class JobScheduler implements Runnable {
   private static final String consumerId = "jobscheduler";
   private static final JsonSerde jsonSerde = new JsonSerde();
 
-  private final ClusterStoreService clusterStoreService;
-  private final ClusterStoreView clusterStore;
+  private final ClusterStore clusterStore;
   private final TrackingQueue provisionerQueue;
   private final TrackingQueue jobQueue;
   private final ZKClient zkClient;
@@ -77,8 +75,7 @@ public class JobScheduler implements Runnable {
                        ZKClient zkClient,
                        TaskService taskService,
                        Configuration conf) {
-    this.clusterStoreService = clusterStoreService;
-    this.clusterStore = clusterStoreService.getView(Account.SYSTEM_ACCOUNT);
+    this.clusterStore = clusterStoreService.getSystemView();
     this.provisionerQueue = provisionerQueue;
     this.jobQueue = jobQueue;
     this.zkClient = ZKClients.namespace(zkClient, Constants.LOCK_NAMESPACE);
@@ -101,7 +98,7 @@ public class JobScheduler implements Runnable {
         ZKInterProcessReentrantLock lock = new ZKInterProcessReentrantLock(zkClient, "/" + jobId.getClusterId());
         try {
           lock.acquire();
-          ClusterJob job = clusterStoreService.getClusterJob(jobId);
+          ClusterJob job = clusterStore.getClusterJob(jobId);
           Cluster cluster = clusterStore.getCluster(job.getClusterId());
           // this can happen if 2 tasks complete around the same time and the first one places the job in the queue,
           // sees 0 in progress tasks, and sets the cluster status. The job is still in the queue as another element
@@ -121,7 +118,7 @@ public class JobScheduler implements Runnable {
           // TODO: avoid looking up every single task every time
           LOG.debug("Verifying task statuses for stage {} for job {}", job.getCurrentStageNumber(), jobIdStr);
           for (String taskId : currentStage) {
-            ClusterTask task = clusterStoreService.getClusterTask(TaskId.fromString(taskId));
+            ClusterTask task = clusterStore.getClusterTask(TaskId.fromString(taskId));
             job.setTaskStatus(task.getTaskId(), task.getStatus());
             LOG.debug("Status of task {} is {}", taskId, task.getStatus());
             if (task.getStatus() == ClusterTask.Status.COMPLETE) {
@@ -171,7 +168,7 @@ public class JobScheduler implements Runnable {
                 taskService.completeJob(job, cluster);
               }
             }
-            clusterStoreService.writeClusterJob(job);
+            clusterStore.writeClusterJob(job);
           } else if (inProgressTasks == 0) {
             // Job failed and no in progress tasks remaining, update cluster status
             taskService.failJobAndSetClusterStatus(job, cluster);
@@ -242,7 +239,7 @@ public class JobScheduler implements Runnable {
 
     // store all retry tasks
     for (ClusterTask t : retryTasks) {
-      clusterStoreService.writeClusterTask(t);
+      clusterStore.writeClusterTask(t);
     }
 
     // Remove self from current stage
@@ -265,7 +262,7 @@ public class JobScheduler implements Runnable {
       return;
     }
 
-    clusterStoreService.writeClusterTask(rollbackTask);
+    clusterStore.writeClusterTask(rollbackTask);
 
     SchedulableTask schedulableTask = new SchedulableTask(rollbackTask);
     LOG.debug("Submitting rollback task {} for task {}", rollbackTask.getTaskId(), task.getTaskId());
